+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 0.18.6
 
 - The LaTeX to PDF forward search now fails in case the pdf is not found (requires TeXiFy 1.1.1).
+- Forward search from TeXiFy now supports projects in WSL (requires TeXiFy 1.1.1).
 
 ## 0.18.5